+++ crates/prelude/CHANGELOG.md
@@ -9,10 +9,12 @@
 
 ### Minor
 
+- Initialize heap for wasm applets
 - Add `scheduling::exit()` for successful applet exit
 
 ### Patch
 
+- Fix rust and clippy lints
 - Make sure at compile-time that at most one `native`, `test`, or `wasm` feature is enabled
 - Update dependencies
 
@@ -172,4 +174,4 @@
 
 ## 0.1.0
 
-<!-- Increment to skip CHANGELOG.md test: 2 -->
+<!-- Increment to skip CHANGELOG.md test: 3 -->