+++ crates/wire-derive/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Patch
 
+- Fix rust and clippy lints
 - Update dependencies
 
 ## 0.1.0